+++ Desktop/GestioneArticoliJspServletJdbc/src/copiaSuGit/it/gestionearticoli/dao/articolo/ArticoloDAO.java
@@ -11,7 +11,7 @@ public interface ArticoloDAO extends IBaseDAO<Articolo> {
 	public TreeSet<Articolo> list() throws Exception;
 	public Articolo get(Long id) throws Exception;
 	public int update(Articolo input) throws Exception;
-	public int insert(Articolo input) throws Exception;
+	public Long insert(Articolo input) throws Exception;
 	public int delete(Articolo input) throws Exception;
 	public TreeSet<Articolo> findByExample(Articolo input) throws Exception;
 	public void setConnection(Connection connection);
